+++ src/base/timeControl.h
@@ -55,7 +55,7 @@ public:
 	void update_real_seconds();
 	float get_vminute_count();
 
-	void notify(const string &task_name, GenericAsyncTask::TaskFunc *function, void* data, int delay = 0.0);
+	void notify(const string &task_name, GenericAsyncTask::TaskFunc *function, void* data, double delay = 0.0);
 	void notify_after_n_frames(int after_n_frames, EventCallbackFunction *function, void *data);
 	void notify_after_n_seconds(int after_n_secs, EventCallbackFunction *function, void *data);
 	void notify_after_n_vminutes(int after_n_vmins, EventCallbackFunction *function, void *data);
